Safari-609.1.11.1/LayoutTests/fast/dom/template-document-dispatchEvent-expected.txt

Test that it is possible to dispatchEvent() works on template documents.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS wasEventDispatched is false
PASS wasEventDispatched is true
PASS successfullyParsed is true

TEST COMPLETE
